dl_modelSelection/requirements.txt
azureml-sdk==1.0.21
azure-common==1.1.26
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==11.0.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.3.0
absl-py==0.11.0
astor==0.8.1
astroid==2.4.2
cached-property==1.5.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.9

